requirements.txt
amqp==5.1.1
aniso8601==9.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bidict==0.22.0
billiard==3.6.4.0
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
cryptography==37.0.2
Deprecated==1.2.13
et-xmlfile==1.1.0
fake-headers==1.0.2
Flask==2.1.2
Flask-Compress==1.12
Flask-Cors==3.0.10
flask-crontab==0.1.2
Flask-JWT-Extended==4.4.2
flask-restplus==0.13.0
flower==1.1.0
greenlet==1.1.2
h11==0.13.0
h5py==3.7.0
html5lib==1.1
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.1
keras==2.9.0
kombu==5.2.4
MarkupSafe==2.1.1
numpy==1.23.0
opencv-python==4.6.0.66
openpyxl==3.0.10
outcome==1.1.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pymongo==4.1.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.3
python-socketio==5.7.0
pytz==2022.1
redis==4.3.4
requests==2.27.1
rq==1.10.1
schedule==1.1.0
scipy==1.8.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
tornado==6.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.4
webencodings==0.5.1
Werkzeug==2.1.2
wrapt==1.14.1
wsproto==1.1.0
xmltodict==0.13.0
zipp==3.8.0
